Extract class to another file
