+++ lib/data/models.dart
@@ -6,30 +6,10 @@ import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 import 'package:rxdart/rxdart.dart';
 
+import 'data_provider.dart';
 import 'firebase/firebase_service.dart';
 import 'models/note_model.dart';
 import 'models/user_model.dart';
-
-@immutable
-class DataProvider {
-  DataProvider._internal();
-  static final DataProvider _instance = DataProvider._internal();
-  factory DataProvider() = DataProvider._internal;
-
-  static UserData<UserModel> get userData => _instance._userData;
-  late final UserData<UserModel> _userData = UserData(
-    converter: (user) => UserModel.fromAuthUser(user),
-  );
-
-  static Collection<NoteModel> get notes => _instance._notes;
-  late final Collection<NoteModel> _notes = Collection<NoteModel>.path(
-    'notes',
-    converter: FirestoreConverter(
-      fromFirestore: (snapshot, options) => NoteModel.fromSnapshot(snapshot),
-      toFirestore: (value, options) => value.toMap(),
-    ),
-  );
-}
 
 class NotesListModel with ChangeNotifier, DiagnosticableTreeMixin {
   NotesListModel({List<NoteModel>? notes}) : _notes = notes ?? [];
